fix: refactor LwM2M format validation process
Refactoring the LwM2M format validation process by introducing a new method: `validateLwM2MFormat`, which replaces the old `checkLwM2MFormat` method. 

The API of the new method is different from the old one. The main difference is that the evaluated object is being returned if validation is successful. 

This change improves the code quality by following the DRY principle in the implementation of the format validation