fix: remove schema from process of remove Coiote format
Update the API of removeCoioteFormatFromSingleInstanceObj by removing one parameter of the method: schema.

the schema was used in the method to check if result follow its definiton. In case the result does not, the value of that resource change to 'undefined'.

However, that decision is not correct because it make not possible to validateLwM2MFormat method check wrong values when resource is not not required.