fix: remove coiote format from 'array instances' objects
the functionality of 'removeCoioteFormatFromArrayInstance' is the same than old method 'convertToLwM2MArrayInstance'

with the difference that the new method do not use the schema parameter to make the conversion.

The validation of the result against the LwM2M Schema definition is happening in a different step ('validateLwM2MFormat' method), for that reason the schema is not needed in the new method.